Remove oauthTokenInfo.go and make common oauth code in one file.
