--- common/oauthTokenInfo.go
@@ -1,57 +1,0 @@
-// Copyright © 2017 Microsoft <wastore@microsoft.com>
-//
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files (the "Software"), to deal
-// in the Software without restriction, including without limitation the rights
-// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-// copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions:
-//
-// The above copyright notice and this permission notice shall be included in
-// all copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-// THE SOFTWARE.
-
-package common
-
-import (
-	"encoding/json"
-
-	"github.com/Azure/go-autorest/autorest/adal"
-)
-
-// OAuthTokenInfo contains info necessary for azcopy to get/refresh OAuth credentials.
-type OAuthTokenInfo struct {
-	adal.Token
-	Tenant                  string `json:"_tenant"`
-	ActiveDirectoryEndpoint string `json:"_ad_endpoint"`
-}
-
-// ToJSON converts OAuthTokenInfo to json format.
-func (credInfo OAuthTokenInfo) ToJSON() ([]byte, error) {
-	return json.Marshal(credInfo)
-}
-
-// IsEmpty returns if current OAuthTokenInfo is empty and doesn't contain any useful info.
-func (credInfo OAuthTokenInfo) IsEmpty() bool {
-	if credInfo.Tenant == "" && credInfo.ActiveDirectoryEndpoint == "" && credInfo.Token.IsZero() {
-		return true
-	}
-
-	return false
-}
-
-// JSONToTokenInfo converts bytes to OAuthTokenInfo
-func JSONToTokenInfo(b []byte) (*OAuthTokenInfo, error) {
-	var OAuthTokenInfo OAuthTokenInfo
-	if err := json.Unmarshal(b, &OAuthTokenInfo); err != nil {
-		return nil, err
-	}
-	return &OAuthTokenInfo, nil
-}
